Add sketch L6474SketchFor1MotorShieldDrivenByUart
Sketch example to drive one motor by commands send via the UART. All the
functions of the library can be accessed.
To use it:
- You have to install the L6474 library that you got from github
- The Uno must have a X-Nucleo-Ihm01A1 on top of it. A stepper motor has
to be connected to the bridges of the expansion board and its external
power supply must be enabled.
- The terminal port must be configured to 9600 baud, new line handled
with option “CR+LF” and I recommend to enable local echo.
- Then, just download the sketch to the Arduino Uno with the IDE.
If your setup is correct, as this step you should see the token :”>>:”
at the serial terminal.
It means that the Uno firmware is expecting some instructions.
Then you have to enter one of the following command line:
C1                -> to get the acceleration
C2                -> to get the current speed
C3                -> to Get the deceleration
C4                -> to Get the shield state
C5                -> to Get the shield the FW version
C6                -> to Get the Mark
C7                -> to Get the max speed state
C8                -> to Get the min speed shield state
C9                -> to Get the current position
C10               -> to go home  the shield state
C11               -> to go to mark the position
C12 <arg1>        -> to go to  <arg1> position
C13               -> to make a hard stop
C14 <arg1> <arg2> -> to move of <arg2> steps in forward direction if
<arg1> is 0,
in backward direction if <arg1> is 1
C15               -> to reset all shields
C16 <arg1>        -> to run in forward direction if <arg1> is 0,
in backward direction if <arg1> is 1
C17 <arg1>        -> to set the acceleration to <arg1>
C18 <arg1>        -> to set the deceleration to <arg1>
C19               -> to set current position to be the home position
C20               -> to set current position to be the mark position
C21 <arg1>        -> to set max speed to <arg1>
C22 <arg1>        -> to set min speed to <arg1>
C23               -> to make a soft stop
C24               -> to wait while the motor is active
C25               -> to send a disable command (bridges off)
C26               -> to send a enable command (bridges on)
C27 <arg1>        -> to send a get param command with  <arg1> as
register address
C28               -> to send a get status command (read and clear status
register)
C29               -> to send a nop command
C30 <arg1> <arg2> -> to send a set param command with  <arg1> as
register address,
<arg2> as value
C31               -> to read status register (no clear is done)
C32               -> to reset the shield
C33               -> to release the reset
C34 <arg1>        -> to set the step mode to full step if <arg1> is 1
to half step if <arg1> is 2
to 1/4 step if <arg1> is 4
to 1/8 step if <arg1> is 8
to 1/16 step if <arg1> is 16
C35 <arg1>        -> to set direction: forward direction if <arg1> is 0,
backward direction if <arg1> is 1